Refactor pointer handling in ft_printf: update puthex and putptr to use uintptr_t for better portability

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dbobrov <dbobrov@student.42wolfsburg.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/24 15:48:04 by dbobrov           #+#    #+#             */
-/*   Updated: 2025/11/29 20:01:30 by dbobrov          ###   ########.fr       */
+/*   Updated: 2025/11/29 20:47:15 by dbobrov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,13 @@
 
 # include "libft/libft.h"
 # include <stdarg.h>
+# include <stdint.h>
 
 size_t	putstr(char *str);
 size_t	putptr(void *ptr);
 size_t	putnbr(int n);
 size_t	putuint(unsigned int n);
-size_t	puthex(unsigned long n, int letter_case);
+size_t	puthex(uintptr_t n, int letter_case);
 int		ft_printf(const char *str, ...);
 size_t	printchar(char c);
